requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=0.21.0

# Social media APIs
atproto>=0.0.31

# Argument parsing and utilities
argparse  # Usually included with Python
